manage user listing

+++ app/Http/Controllers/ListingController.php
@@ -53,6 +53,10 @@ class ListingController extends Controller
     }
 
     public function update(Request $request,Listing $listing){
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized Action');
+        }
         $formFields = $request->validate([
             'title'=>'required',
             'company'=>'required',
@@ -72,7 +76,14 @@ class ListingController extends Controller
         return back()->with('message','listing updated successfully!');
     }
     public function destroy(Listing $listing){
+        //make sure logged in user is owner
+        if($listing->user_id != auth()->id()){
+            abort(403,'Unauthorized Action');
+        }
         $listing->delete();
         return redirect('/')->with('message','listing deleted successfully!');
+    }
+    public function manage(){
+        return view('listings.manage',['listings'=> auth()->user()->listings()->get()]);
     }
 }
